Moved entities to domain/model
